pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yt-finance-analyzer"
version = "0.1.0"
description = "YouTube 財經影片自動分析系統"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "yt-dlp>=2025.1.0",
    "google-api-python-client>=2.100.0",
    "openai>=1.6.0",
    "anthropic>=0.39.0",
    "Jinja2>=3.1.2",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]

[project.scripts]
yt-finance-analyzer = "yt_finance_analyzer.main:main"

[tool.setuptools.packages.find]
where = ["src"]
